requirements.txt
google-adk==1.21.0
mcp==1.10.0
pydantic==2.12.5
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0